refactor(gantt): split the chart component into feature hooks and layers
The chart component had grown to roughly 1,100 lines of body. The row model,
the viewport, keyboard navigation, scheduling, selection and the task list
pane all lived in one function, so reading any one of them meant scrolling
past the other five, and a change to one sat next to state it had nothing to
do with.

Each of those now owns a hook, and the render owns three presentational
layers. Nothing about the public API changes: the generated declarations are
identical apart from the module paths, and the emitted CSS is byte for byte
the same.

- `useGanttRowModel` - hierarchy, collapsing and grouping into the one `rows`
  array both panes read
- `useGanttViewport` - the timeline rebuild and its scroll compensation, the
  range extension, wheel zoom, range reporting and the initial scroll
- `useGanttKeyboardNav` - the roving tabindex, the edit shortcuts and the
  live-region announcements
- `useGanttScheduling` - the one "is this day off" answer, the working
  calendar, the propagation policy and the critical path
- `useGanttCollapse`, `useGanttSelection`, `useGanttTaskListPane`,
  `useGanttInteraction` - the smaller pieces of state, each with its own
  invariant
- `GanttToolbar`, `GanttRowsLayer`, `GanttBarsLayer`,
  `GanttNonWorkingLayer` - the render, one layer at a time

Along the way:

- `src/pages/Gantt.tsx` moves to `src/components/Gantt/`. The library has no
  pages, and the name said otherwise.
- `GanttProps` moves next to it, so the component file is wiring rather than
  330 lines of prop documentation.
- `isRowExpandable` replaces the same expression written out in the keyboard
  model and again in the render - the twisty and the key that operates it
  cannot disagree now.
- `useLatestRef` replaces the block of six refs the effects synced by hand.
- `useGanttSelectors` keeps only what the component itself renders with;
  every hook subscribes to the store actions it needs.
