+++ docs/en/ref/interaction-config.md
@@ -212,7 +212,7 @@ export function canCreateTasks(
 with the presence of the callback:
 
 ```ts
-// src/pages/Gantt.tsx
+// src/components/Gantt/Gantt.tsx
 const canDrawTasks = onTaskCreate !== undefined && canCreateTasks(interaction);
 ```
