+++ docs/en/ref/markers.md
@@ -109,7 +109,7 @@ A task ending exactly on `date` is not an overrun.
 The chart prepends one marker of its own to whatever `markers` holds:
 
 ```ts
-// src/pages/Gantt.tsx
+// src/components/Gantt/Gantt.tsx
 { id: "today", date: dayjs(), className: "gantt-today-marker" }
 ```
