+++ docs/ko/ref/interaction-config.md
@@ -209,7 +209,7 @@ export function canCreateTasks(
 AND로 묶어요:
 
 ```ts
-// src/pages/Gantt.tsx
+// src/components/Gantt/Gantt.tsx
 const canDrawTasks = onTaskCreate !== undefined && canCreateTasks(interaction);
 ```
