+++ docs/ko/ref/markers.md
@@ -109,7 +109,7 @@ const overrun =
 차트는 `markers`에 무엇이 담겨 있든 자기 마커 하나를 앞에 덧붙여요:
 
 ```ts
-// src/pages/Gantt.tsx
+// src/components/Gantt/Gantt.tsx
 { id: "today", date: dayjs(), className: "gantt-today-marker" }
 ```
