+++ eslint.config.js
@@ -89,7 +89,6 @@ export default tseslint.config(
                 "@tanstack/*",
                 "components/*",
                 "hooks/*",
-                "pages/*",
                 "stores/*",
                 "constants/*",
                 "types/*",
@@ -97,7 +96,6 @@ export default tseslint.config(
                 "assets/*",
                 "**/components/*",
                 "**/hooks/*",
-                "**/pages/*",
                 "**/stores/*",
                 "**/constants/*",
                 "**/utils/*",
